:recycle: Move styles to Site, remove theme readme
Move styles to Site, remove theme readme
