+++ gatsby-theme-fela/src/pages/index.js
@@ -3,7 +3,6 @@ import { Link } from "gatsby"
 import { connect as connectFela } from "react-fela"
 
 import Layout from "../components/Layout"
-import Image from "../components/Image"
 import SEO from "../components/Seo"
 
 export const imageSection = () => ({
@@ -17,9 +16,6 @@ const IndexPage = ({ styles }) => (
     <h1>Hi people</h1>
     <p>Welcome to your new Gatsby site.</p>
     <p>Now go build something great.</p>
-    <div className={styles.imageSection}>
-      <Image />
-    </div>
     <Link to="/page-2/">Go to page 2</Link>
   </Layout>
 )
